Move marketplace abstraction into separate file
